requirements.txt
torch
torchvision
torchaudio
torch-geometric
scanpy
squidpy
anndata
scikit-learn
matplotlib
seaborn
pyyaml
tqdm
